build-agent: /chore: move types from core to types directory
